Move group dialogs to window_dialogs.py, rename logic to GroupManager
Create/edit-group, rename-tag, and the move/copy group-assignment dialogs
now live in WindowConfigDialogsMixin; the near-duplicate create/edit forms
share one _group_form_dialog builder. GroupManager gains rename_group()
(the edit dialog previously mutated group_info in place and relied on
set_group_color to persist). Also deletes the stale shadowed copies of
these actions from actions.py (MainWindow's own definitions had superseded
them).
