+++ sshpilot/actions.py
@@ -3,7 +3,6 @@
 import logging
 import os
 import random
-from typing import Optional
 from gi.repository import Gio, Gtk, Adw, GLib, Gdk
 from gettext import gettext as _
 
@@ -436,173 +435,6 @@ class WindowActions:
         except Exception as e:
             logger.error(f"Failed to open known hosts editor: {e}")
 
-    def on_create_group_action(self, action, param=None):
-        """Handle create group action"""
-        try:
-            dialog = Adw.MessageDialog(
-                transient_for=self,
-                modal=True,
-                heading=_('Create Group'),
-                body=_('Enter a name for the new group:'),
-            )
-
-            content_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=12)
-            content_box.set_margin_start(20)
-            content_box.set_margin_end(20)
-            content_box.set_margin_top(20)
-            content_box.set_margin_bottom(20)
-
-            entry = Gtk.Entry()
-            entry.set_placeholder_text(_('e.g., Work Servers'))
-            entry.set_activates_default(True)
-            entry.set_hexpand(True)
-            content_box.append(entry)
-
-            color_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-            color_row.set_hexpand(True)
-            color_label = Gtk.Label(label=_("Group color"))
-            color_label.set_xalign(0)
-            color_label.set_hexpand(True)
-            color_row.append(color_label)
-
-            color_controls = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-            color_button = Gtk.ColorButton()
-            color_button.set_use_alpha(True)
-            color_button.set_title(_("Select group color"))
-            default_rgba = Gdk.RGBA()
-            default_rgba.red = default_rgba.green = default_rgba.blue = 0
-            default_rgba.alpha = 0
-            color_button.set_rgba(default_rgba)
-            color_controls.append(color_button)
-
-            color_selected = False
-
-            def mark_color_selected(_button):
-                nonlocal color_selected
-                color_selected = True
-
-            color_button.connect('color-set', mark_color_selected)
-
-            def reset_color_selection() -> None:
-                nonlocal color_selected
-                color_selected = False
-                cleared = Gdk.RGBA()
-                cleared.red = cleared.green = cleared.blue = 0
-                cleared.alpha = 0
-                color_button.set_rgba(cleared)
-
-            clear_color_button = Gtk.Button(label=_("Clear"))
-            clear_color_button.add_css_class('flat')
-            clear_color_button.connect('clicked', lambda _btn: reset_color_selection())
-            color_controls.append(clear_color_button)
-
-            color_row.append(color_controls)
-            content_box.append(color_row)
-
-            dialog.set_extra_child(content_box)
-
-            dialog.add_response('cancel', _('Cancel'))
-            dialog.add_response('create', _('Create'))
-            dialog.set_response_appearance('create', Adw.ResponseAppearance.SUGGESTED)
-            dialog.set_default_response('create')
-            dialog.set_close_response('cancel')
-
-            def on_response(dialog, response):
-                if response == 'create':
-                    name = entry.get_text().strip()
-                    if name:
-                        selected_color = None
-                        rgba_value = color_button.get_rgba()
-                        if color_selected and rgba_value.alpha > 0:
-                            selected_color = rgba_value.to_string()
-                        self.group_manager.create_group(name, color=selected_color)
-                        self.rebuild_connection_list()
-                    else:
-                        error_dialog = Adw.MessageDialog(
-                            transient_for=self,
-                            modal=True,
-                            heading=_("Error"),
-                            body=_("Please enter a group name."),
-                        )
-                        error_dialog.add_response('ok', _('OK'))
-                        error_dialog.present()
-                dialog.destroy()
-
-            dialog.connect('response', on_response)
-            dialog.present()
-
-            def focus_entry():
-                entry.grab_focus()
-                return False
-
-            GLib.idle_add(focus_entry)
-
-        except Exception as e:
-            logger.error(f"Failed to show create group dialog: {e}")
-
-    def on_edit_group_action(self, action, param=None):
-        """Handle edit group action"""
-        try:
-            selected_row = getattr(self, '_context_menu_group_row', None)
-            if not selected_row:
-                selected_row = self.connection_list.get_selected_row()
-            if not selected_row or not hasattr(selected_row, 'group_id'):
-                return
-
-            group_id = selected_row.group_id
-            group_info = self.group_manager.groups.get(group_id)
-            if not group_info:
-                return
-
-            dialog = Adw.MessageDialog(
-                transient_for=self,
-                modal=True,
-                heading=_('Edit Group'),
-                body=_('Enter a new name for the group:'),
-            )
-
-            entry = Gtk.Entry(text=group_info['name'])
-            entry.set_activates_default(True)
-            entry.set_hexpand(True)
-            dialog.set_extra_child(entry)
-
-            dialog.add_response('cancel', _('Cancel'))
-            dialog.add_response('save', _('Save'))
-            dialog.set_response_appearance('save', Adw.ResponseAppearance.SUGGESTED)
-            dialog.set_default_response('save')
-            dialog.set_close_response('cancel')
-
-            def on_response(dialog, response):
-                if response == 'save':
-                    new_name = entry.get_text().strip()
-                    if new_name:
-                        group_info['name'] = new_name
-                        self.group_manager._save_groups()
-                        self.rebuild_connection_list()
-                    else:
-                        error_dialog = Adw.MessageDialog(
-                            transient_for=self,
-                            modal=True,
-                            heading=_("Error"),
-                            body=_("Please enter a group name."),
-                        )
-                        error_dialog.add_response('ok', _('OK'))
-                        error_dialog.present()
-                dialog.destroy()
-
-            dialog.connect('response', on_response)
-            dialog.present()
-
-            def focus_entry():
-                entry.grab_focus()
-                entry.select_region(0, -1)
-                return False
-
-            GLib.idle_add(focus_entry)
-
-        except Exception as e:
-            logger.error(f"Failed to show edit group dialog: {e}")
-
     def on_delete_group_action(self, action, param=None):
         """Handle delete group action"""
         try:
@@ -692,28 +524,6 @@ class WindowActions:
 
         except Exception as e:
             logger.error(f"Failed to show delete group dialog: {e}")
-
-    def on_move_to_ungrouped_action(self, action, param=None):
-        """Handle move to ungrouped action"""
-        try:
-            # Get the connection from context menu or selected row
-            selected_row = getattr(self, '_context_menu_connection', None)
-            if not selected_row:
-                selected_row = self.connection_list.get_selected_row()
-                if selected_row and hasattr(selected_row, 'connection'):
-                    selected_row = selected_row.connection
-
-            if not selected_row:
-                return
-
-            connection_nickname = selected_row.nickname if hasattr(selected_row, 'nickname') else selected_row
-
-            # Move to ungrouped (None group)
-            self.group_manager.move_connection(connection_nickname, None)
-            self.rebuild_connection_list()
-
-        except Exception as e:
-            logger.error(f"Failed to move connection to ungrouped: {e}")
 
     def on_export_config_action(self, action, param=None):
         """Handle export configuration action"""
@@ -1074,246 +884,6 @@ class WindowActions:
                 self.welcome_view.refresh_pinned()
         except Exception as exc:
             logger.debug(f"Failed to refresh pinned sessions: {exc}")
-
-    def on_move_to_group_action(self, action, param=None):
-        """Handle move to group action"""
-        try:
-            # Get the connection from context menu or selected row
-            selected_row = getattr(self, '_context_menu_connection', None)
-            if not selected_row:
-                selected_row = self.connection_list.get_selected_row()
-                if selected_row and hasattr(selected_row, 'connection'):
-                    selected_row = selected_row.connection
-
-            if not selected_row:
-                return
-
-            connection_nickname = selected_row.nickname if hasattr(selected_row, 'nickname') else selected_row
-
-            # Get available groups
-            available_groups = self.get_available_groups()
-            logger.debug(f"Available groups for move dialog: {len(available_groups)} groups")
-
-            dialog = Adw.MessageDialog(
-                transient_for=self,
-                modal=True,
-                heading=_("Move to Group"),
-                body=_("Select a group to move the connection to:"),
-            )
-
-            content_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=12)
-            content_box.set_margin_start(20)
-            content_box.set_margin_end(20)
-            content_box.set_margin_top(20)
-            content_box.set_margin_bottom(20)
-
-            listbox = Gtk.ListBox()
-            listbox.set_selection_mode(Gtk.SelectionMode.SINGLE)
-            listbox.set_vexpand(True)
-
-            create_section_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=6)
-            create_section_box.set_margin_start(12)
-            create_section_box.set_margin_end(12)
-            create_section_box.set_margin_top(6)
-            create_section_box.set_margin_bottom(6)
-
-            create_label = Gtk.Label(label=_("Create New Group"))
-            create_label.set_xalign(0)
-            create_label.add_css_class("heading")
-            create_section_box.append(create_label)
-
-            create_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-
-            create_group_entry = Gtk.Entry()
-            create_group_entry.set_placeholder_text(_("Enter group name"))
-            create_group_entry.set_hexpand(True)
-            create_group_entry.set_activates_default(True)
-            create_box.append(create_group_entry)
-
-            create_section_box.append(create_box)
-
-            color_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-            color_row.set_hexpand(True)
-            color_label = Gtk.Label(label=_("Group color"))
-            color_label.set_xalign(0)
-            color_label.set_hexpand(True)
-            color_row.append(color_label)
-
-            color_controls = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-            color_button = Gtk.ColorButton()
-            color_button.set_use_alpha(True)
-            color_button.set_title(_("Select group color"))
-            initial_rgba = Gdk.RGBA()
-            initial_rgba.red = initial_rgba.green = initial_rgba.blue = 0
-            initial_rgba.alpha = 0
-            color_button.set_rgba(initial_rgba)
-            color_controls.append(color_button)
-
-            color_selected = False
-
-            def mark_color_selected(_button):
-                nonlocal color_selected
-                color_selected = True
-
-            color_button.connect('color-set', mark_color_selected)
-
-            def reset_color_selection() -> None:
-                nonlocal color_selected
-                color_selected = False
-                cleared = Gdk.RGBA()
-                cleared.red = cleared.green = cleared.blue = 0
-                cleared.alpha = 0
-                color_button.set_rgba(cleared)
-
-            clear_color_button = Gtk.Button(label=_("Clear"))
-            clear_color_button.add_css_class('flat')
-            clear_color_button.connect('clicked', lambda _btn: reset_color_selection())
-            color_controls.append(clear_color_button)
-
-            color_row.append(color_controls)
-            create_section_box.append(color_row)
-            content_box.append(create_section_box)
-
-            separator = Gtk.Separator(orientation=Gtk.Orientation.HORIZONTAL)
-            content_box.append(separator)
-
-            if available_groups:
-                existing_label = Gtk.Label(label=_("Existing Groups"))
-                existing_label.set_xalign(0)
-                existing_label.add_css_class("heading")
-                content_box.append(existing_label)
-
-            for group in available_groups:
-                row = Gtk.ListBoxRow()
-                box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-
-                from sshpilot import icon_utils
-                icon = icon_utils.new_image_from_icon_name('folder-symbolic')
-                icon.set_pixel_size(16)
-                box.append(icon)
-
-                name_label = Gtk.Label(label=group['name'])
-                name_label.set_xalign(0)
-                name_label.set_hexpand(True)
-                box.append(name_label)
-
-                row.set_child(box)
-                row.group_id = group['id']
-                listbox.append(row)
-
-            content_box.append(listbox)
-
-            dialog.set_extra_child(content_box)
-
-            dialog.add_response('cancel', _('Cancel'))
-            dialog.add_response('move', _('Move'))
-            dialog.set_response_appearance('move', Adw.ResponseAppearance.SUGGESTED)
-            dialog.set_default_response('move')
-            dialog.set_close_response('cancel')
-
-            # Connect entry and button events
-            def find_existing_group_id(name: str) -> Optional[str]:
-                lowered = name.lower()
-                for group in available_groups:
-                    if group['name'].lower() == lowered:
-                        return group['id']
-                return None
-
-            def has_valid_target() -> bool:
-                if create_group_entry.get_text().strip():
-                    return True
-                return listbox.get_selected_row() is not None
-
-            def update_move_response_state(*_args) -> None:
-                dialog.set_response_enabled('move', has_valid_target())
-
-            listbox.connect('row-selected', lambda _lb, _row: update_move_response_state())
-
-            def on_entry_changed(_entry):
-                update_move_response_state()
-
-            def on_entry_activated(_entry):
-                if has_valid_target():
-                    dialog.response('move')
-
-            create_group_entry.connect('changed', on_entry_changed)
-            create_group_entry.connect('activate', on_entry_activated)
-            update_move_response_state()
-
-            def perform_move() -> bool:
-                group_name = create_group_entry.get_text().strip()
-                if group_name:
-                    existing_group_id = find_existing_group_id(group_name)
-                    if existing_group_id:
-                        self.group_manager.move_connection(connection_nickname, existing_group_id)
-                        self.rebuild_connection_list()
-                        return True
-                    try:
-                        selected_color = None
-                        rgba_value = color_button.get_rgba()
-                        if color_selected and rgba_value.alpha > 0:
-                            selected_color = rgba_value.to_string()
-                        new_group_id = self.group_manager.create_group(group_name, color=selected_color)
-                        self.group_manager.move_connection(connection_nickname, new_group_id)
-                        self.rebuild_connection_list()
-                        return True
-                    except ValueError as e:
-                        error_dialog = Gtk.Dialog(
-                            title=_("Group Already Exists"),
-                            transient_for=dialog,
-                            modal=True,
-                            destroy_with_parent=True
-                        )
-                        error_dialog.set_default_size(400, 150)
-                        error_dialog.set_resizable(False)
-
-                        content_area = error_dialog.get_content_area()
-                        content_area.set_margin_start(20)
-                        content_area.set_margin_end(20)
-                        content_area.set_margin_top(20)
-                        content_area.set_margin_bottom(20)
-
-                        error_label = Gtk.Label(label=str(e))
-                        error_label.set_wrap(True)
-                        error_label.set_xalign(0)
-                        content_area.append(error_label)
-
-                        error_dialog.add_button(_('OK'), Gtk.ResponseType.OK)
-                        error_dialog.set_default_response(Gtk.ResponseType.OK)
-
-                        def on_error_response(dialog, _response):
-                            dialog.destroy()
-
-                        error_dialog.connect('response', on_error_response)
-                        error_dialog.present()
-
-                        create_group_entry.set_text("")
-                        reset_color_selection()
-                        create_group_entry.grab_focus()
-                        update_move_response_state()
-                        return False
-
-                selected_row = listbox.get_selected_row()
-                if selected_row:
-                    target_group_id = selected_row.group_id
-                    self.group_manager.move_connection(connection_nickname, target_group_id)
-                    self.rebuild_connection_list()
-                    return True
-                return False
-
-            def on_response(dialog, response):
-                if response == 'move':
-                    if perform_move():
-                        dialog.destroy()
-                        return
-                    return
-                dialog.destroy()
-
-            dialog.connect('response', on_response)
-            dialog.present()
-
-        except Exception as e:
-            logger.error(f"Failed to show move to group dialog: {e}")
 
     def on_check_for_updates_action(self, action, param=None):
         """Handle check for updates action from menu"""
